requirements.txt
# Core FastAPI and WebSocket dependencies
fastapi>=0.100.0,<0.105.0
uvicorn[standard]>=0.20.0,<0.30.0
websockets>=10.0,<13.0
python-multipart>=0.0.6,<0.1.0

# Database dependencies
sqlalchemy>=2.0.0,<3.0.0
aiosqlite>=0.19.0,<0.20.0

# ADK dependencies
google-adk>=0.13.0,<0.15.0
google-adk[vertexai]>=0.13.0,<0.15.0

# Data validation and serialization
pydantic>=2.0.0,<3.0.0

# CORS and middleware
starlette>=0.27.0,<0.30.0

# Logging and utilities
python-dotenv>=1.0.0,<2.0.0
PyYAML>=6.0.0,<7.0.0

# Development dependencies (optional)
pytest>=7.0.0,<8.0.0
pytest-asyncio>=0.21.0,<0.22.0
httpx>=0.24.0,<0.26.0

# Additional dependencies for DevOps tools
elasticsearch>=7.0.0,<9.0.0
requests>=2.31.0,<3.0.0
kubernetes>=28.1.0,<29.0.0